Struct x509_certificate::certificate::CapturedX509Certificate

pub struct CapturedX509Certificate { /* private fields */ }

Represents an immutable (read-only) X.509 certificate that was parsed from data.

This type implements Deref but not DerefMut, so only functions taking a non-mutable instance are usable.

A copy of the certificate’s raw backing data is stored, facilitating subsequent access.

Implementations

impl CapturedX509Certificate

pub fn from_der(data: impl Into<Vec<u8>>) -> Result<Self, Error>

Construct an instance from DER encoded data.

A copy of this data will be stored in the instance and is guaranteed to be immutable for the lifetime of the instance. The original constructing data can be retrieved later.

pub fn from_ber(data: impl Into<Vec<u8>>) -> Result<Self, Error>

Construct an instance from BER encoded data.

A copy of this data will be stored in the instance and is guaranteed to be immutable for the lifetime of the instance, allowing it to be retrieved later.

pub fn from_pem(data: impl AsRef<[u8]>) -> Result<Self, Error>

Construct an instance by parsing PEM encoded ASN.1 data.

The data is a human readable string likely containing --------- BEGIN CERTIFICATE ----------.

pub fn from_pem_multiple(data: impl AsRef<[u8]>) -> Result<Vec<Self>, Error>

Construct instances by parsing PEM with potentially multiple records.

By default, we only look for --------- BEGIN CERTIFICATE -------- entries and silently ignore unknown ones. If you would like to specify an alternate set of tags (this is the value after the BEGIN) to search, call Self::from_pem_multiple_tags.

pub fn from_pem_multiple_tags( data: impl AsRef<[u8]>, tags: &[&str] ) -> Result<Vec<Self>, Error>

Construct instances by parsing PEM armored DER encoded certificates with specific PEM tags.

This is like Self::from_pem_multiple except you control the filter for which BEGIN <tag> values are filtered through to the DER parser.

pub fn constructed_data(&self) -> &[u8]

Obtain the DER data that was used to construct this instance.

The data is guaranteed to not have been modified since the instance was constructed.

pub fn encode_pem(&self) -> String

Encode the original contents of this certificate to PEM.

pub fn verify_signed_by_certificate( &self, other: impl AsRef<X509Certificate> ) -> Result<(), Error>

Verify that another certificate, other, signed this certificate.

If this is a self-signed certificate, you can pass self as the 2nd argument.

This function isn’t exposed on X509Certificate because the exact bytes constituting the certificate’s internals need to be consulted to verify signatures. And since this type tracks the underlying bytes, we are guaranteed to have a pristine copy.

pub fn verify_signed_data( &self, signed_data: impl AsRef<[u8]>, signature: impl AsRef<[u8]> ) -> Result<(), Error>

Verify a signature over signed data purportedly signed by this certificate.

This is a wrapper to Self::verify_signed_data_with_algorithm() that will derive the verification algorithm from the public key type type and the signature algorithm indicated in this certificate. Typically these align. However, it is possible for a signature to be produced with a different digest algorithm from that indicated in this certificate.

pub fn verify_signed_data_with_algorithm( &self, signed_data: impl AsRef<[u8]>, signature: impl AsRef<[u8]>, verify_algorithm: &'static dyn VerificationAlgorithm ) -> Result<(), Error>

Verify a signature over signed data using an explicit verification algorithm.

This is like Self::verify_signed_data() except the verification algorithm to use is passed in instead of derived from the default algorithm for the signing key’s type.

pub fn verify_signed_by_public_key( &self, public_key_data: impl AsRef<[u8]> ) -> Result<(), Error>

Verifies that this certificate was cryptographically signed using raw public key data from a signing key.

This function does the low-level work of extracting the signature and verification details from the current certificate and figuring out the correct combination of cryptography settings to apply to perform signature verification.

In many cases, an X.509 certificate is signed by another certificate. And since the public key is embedded in the X.509 certificate, it is easier to go through Self::verify_signed_by_certificate instead.

pub fn find_signing_certificate<'a>( &self, certs: impl Iterator<Item = &'a Self> ) -> Option<&'a Self>

Attempt to find the issuing certificate of this one.

Given an iterable of certificates, we find the first certificate where we are able to verify that our signature was made by their public key.

This function can yield false negatives for cases where we don’t support the signature algorithm on the incoming certificates.

pub fn resolve_signing_chain<'a>( &self, certs: impl Iterator<Item = &'a Self> ) -> Vec<&'a Self>

Attempt to resolve the signing chain of this certificate.

Given an iterable of certificates, we recursively resolve the chain of certificates that signed this one until we are no longer able to find any more certificates in the input set.

Like Self::find_signing_certificate, this can yield false negatives (read: an incomplete chain) due to run-time failures, such as lack of support for a certificate’s signature algorithm.

As a certificate is encountered, it is removed from the set of future candidates.

The traversal ends when we get to an identical certificate (its DER data is equivalent) or we couldn’t find a certificate in the remaining set that signed the last one.

Because we need to recursively verify certificates, the incoming iterator is buffered.