Struct x509_certificate::certificate::X509Certificate

pub struct X509Certificate(_);

Provides an interface to the RFC 5280 rfc5280::Certificate ASN.1 type.

This type provides the main high-level API that this crate exposes for reading and writing X.509 certificates.

Instances are backed by an actual ASN.1 rfc5280::Certificate instance. Read operations are performed against the raw ASN.1 values. Mutations result in mutations of the ASN.1 data structures.

Instances can be converted to/from rfc5280::Certificate using traits. AsRef/AsMut are implemented to obtain a reference to the backing rfc5280::Certificate.

We have chosen not to implement Deref/DerefMut because we don’t want to pollute the type’s API with lower-level ASN.1 primitives.

This type does not track the original data from which it came. If you want a type that does that, consider CapturedX509Certificate, which implements Deref and therefore behaves like this type.

Implementations

impl X509Certificate

pub fn from_der(data: impl AsRef<[u8]>) -> Result<Self, Error>

Construct an instance by parsing DER encoded ASN.1 data.

pub fn from_ber(data: impl AsRef<[u8]>) -> Result<Self, Error>

Construct an instance by parsing BER encoded ASN.1 data.

X.509 certificates are likely (and should be) using DER encoding. However, some specifications do mandate the use of BER, so this method is provided.

pub fn from_pem(data: impl AsRef<[u8]>) -> Result<Self, Error>

Construct an instance by parsing PEM encoded ASN.1 data.

The data is a human readable string likely containing --------- BEGIN CERTIFICATE ----------.

pub fn from_pem_multiple(data: impl AsRef<[u8]>) -> Result<Vec<Self>, Error>

Construct instances by parsing PEM with potentially multiple records.

By default, we only look for --------- BEGIN CERTIFICATE -------- entries and silently ignore unknown ones. If you would like to specify an alternate set of tags (this is the value after the BEGIN) to search, call Self::from_pem_multiple_tags.

pub fn from_pem_multiple_tags( data: impl AsRef<[u8]>, tags: &[&str] ) -> Result<Vec<Self>, Error>

Construct instances by parsing PEM armored DER encoded certificates with specific PEM tags.

This is like Self::from_pem_multiple except you control the filter for which BEGIN <tag> values are filtered through to the DER parser.

pub fn serial_number_asn1(&self) -> &Integer

Obtain the serial number as the ASN.1 Integer type.

pub fn subject_name(&self) -> &Name

Obtain the certificate’s subject, as its ASN.1 Name type.

pub fn subject_common_name(&self) -> Option<String>

Obtain the Common Name (CN) attribute from the certificate’s subject, if set and decodable.

pub fn issuer_name(&self) -> &Name

Obtain the certificate’s issuer, as its ASN.1 Name type.

pub fn issuer_common_name(&self) -> Option<String>

Obtain the Common Name (CN) attribute from the certificate’s issuer, if set and decodable.

pub fn iter_extensions(&self) -> impl Iterator<Item = &Extension>

Iterate over extensions defined in this certificate.

pub fn encode_der_to(&self, fh: &mut impl Write) -> Result<(), Error>

Encode the certificate data structure using DER encoding.

(This is the common ASN.1 encoding format for X.509 certificates.)

This always serializes the internal ASN.1 data structure. If you call this on a wrapper type that has retained a copy of the original data, this may emit different data than that copy.

pub fn encode_ber_to(&self, fh: &mut impl Write) -> Result<(), Error>

Encode the certificate data structure use BER encoding.

pub fn encode_der(&self) -> Result<Vec<u8>, Error>

Encode the internal ASN.1 data structures to DER.

pub fn encode_ber(&self) -> Result<Vec<u8>, Error>

Obtain the BER encoded representation of this certificate.

pub fn write_pem(&self, fh: &mut impl Write) -> Result<(), Error>

Encode the certificate to PEM.

This will write a human-readable string with ------ BEGIN CERTIFICATE ------- armoring. This is a very common method for encoding certificates.

The underlying binary data is DER encoded.

pub fn encode_pem(&self) -> Result<String, Error>

Encode the certificate to a PEM string.

pub fn key_algorithm(&self) -> Option<KeyAlgorithm>

Attempt to resolve a known KeyAlgorithm used by the private key associated with this certificate.

If this crate isn’t aware of the OID associated with the key algorithm, None is returned.

pub fn key_algorithm_oid(&self) -> &Oid

Obtain the OID of the private key’s algorithm.

pub fn signature_algorithm(&self) -> Option<SignatureAlgorithm>

Obtain the [SignatureAlgorithm this certificate will use.

Returns None if we failed to resolve an instance (probably because we don’t recognize the algorithm).

pub fn signature_algorithm_oid(&self) -> &Oid

Obtain the OID of the signature algorithm this certificate will use.

pub fn signature_signature_algorithm(&self) -> Option<SignatureAlgorithm>

Obtain the SignatureAlgorithm used to sign this certificate.

Returns None if we failed to resolve an instance (probably because we don’t recognize that algorithm).

pub fn signature_signature_algorithm_oid(&self) -> &Oid

Obtain the OID of the signature algorithm used to sign this certificate.

pub fn public_key_data(&self) -> Bytes

Obtain the raw data constituting this certificate’s public key.

A copy of the data is returned.

pub fn rsa_public_key_data(&self) -> Result<RsaPublicKey, Error>

Attempt to parse the public key data as RsaPublicKey parameters.

Note that the raw integer value for modulus has a leading 0 byte. So its raw length will be 1 greater than key length. e.g. an RSA 2048 key will have value.modulus.as_slice().len() == 257 instead of 256.

pub fn compare_issuer(&self, other: &Self) -> Ordering

Compare 2 instances, sorting them so the issuer comes before the issued.

This function examines the Self::issuer_name and Self::subject_name fields of 2 certificates, attempting to sort them so the issuing certificate comes before the issued certificate.

This function performs a strict compare of the ASN.1 Name data. The assumption here is that the issuing certificate’s subject Name is identical to the issued’s issuer Name. This assumption is often true. But it likely isn’t always true, so this function may not produce reliable results.

pub fn subject_is_issuer(&self) -> bool

Whether the subject Name is also the issuer’s Name.

This might be a way of determining if a certificate is self-signed. But there can likely be false negatives due to differences in ASN.1 encoding of the underlying data. So we don’t claim this is a test for being self-signed.

pub fn fingerprint(&self, algorithm: DigestAlgorithm) -> Result<Digest, Error>

Obtain the fingerprint for this certificate given a digest algorithm.

pub fn sha1_fingerprint(&self) -> Result<Digest, Error>

Obtain the SHA-1 fingerprint of this certificate.

pub fn sha256_fingerprint(&self) -> Result<Digest, Error>

Obtain the SHA-256 fingerprint of this certificate.