Enum x509_certificate::signing::InMemorySigningKeyPair

pub enum InMemorySigningKeyPair {
    Ecdsa(EcdsaKeyPair, EcdsaCurve, Vec<u8>),
    Ed25519(Ed25519KeyPair),
    Rsa(RsaKeyPair, Vec<u8>),
}

Represents a key pair that exists in memory and can be used to create cryptographic signatures.

This is a wrapper around ring’s various key pair types. It provides abstractions tailored for X.509 certificates.

Variants

Ecdsa(EcdsaKeyPair, EcdsaCurve, Vec<u8>)

ECDSA key pair.

Ed25519(Ed25519KeyPair)

ED25519 key pair.

Rsa(RsaKeyPair, Vec<u8>)

RSA key pair.

Implementations

impl InMemorySigningKeyPair

pub fn from_pkcs8_der(data: impl AsRef<[u8]>) -> Result<Self, Error>

Attempt to instantiate an instance from PKCS#8 DER data.

The DER data should be a OneAsymmetricKey ASN.1 structure.

pub fn from_pkcs8_pem(data: impl AsRef<[u8]>) -> Result<Self, Error>

Attempt to instantiate an instance from PEM encoded PKCS#8.

This is just a wrapper for Self::from_pkcs8_der that does the PEM decoding for you.

pub fn generate_random( key_algorithm: KeyAlgorithm ) -> Result<(Self, Document), Error>

Generate a random key pair given a key algorithm and optional ECDSA signing algorithm.

The raw PKCS#8 document is returned to facilitate access to the private key.

Not attempt is made to protect the private key in memory.

pub fn verification_algorithm( &self ) -> Result<&'static dyn VerificationAlgorithm, Error>

Attempt to resolve a verification algorithm for this key pair.

This is a wrapper around SignatureAlgorithm::resolve_verification_algorithm() with our bound KeyAlgorithm. However, since there are no parameters that can result in wrong choices, this is guaranteed to always work and doesn’t require Result.